Substitute url_w_credentials for url_w_subdir
What?
=====

Uses `url_w_credentials` in place of `url_w_subdir` when looking up
channels in the priority map.

Why?
====

We use a private channel repository in conda and we are getting an
error here.
Stepping through the source code in a debugger it looks like the dict
key values of `_channel_priority_map` match the values of
`url_w_credentials` for cases where we are logged in (and do not need
 more than the channel name) and when we are logged out of anaconda
 (and thus require a url + token as our channel).