micromamba: add another `run` name generation failsafe  (#1491)
When generating `<prefix>_<program>` names,
we might end up with no more prefixes available (because
other programs are running with the same name).
In this case we try to change the name and re-try with the various prefixes.
If finally we can't find an available name, we just generate a prefix randomly.