micromamba: remove explicit progress bar termination
Previously, explicitly terminating the progress bar was necessary
to avoid the progress bars threads
from running while reporting
erros in `main()` thread.
The issue was due to no `.terminate()` call happening in progress bars threads
when an exception was thrown.

We now rely on `interrupt_guard` to call the progress bars termination
when an exception  escapes it's scope, already used in progress bar threads.

The previous explicit fix is therefore unnecessary.