Fixed error when attempting to get the size of directories.
The standard filesystem library cannot guarantee
that attempting to get the size
of a directory with result in a value or an error,
because of the wild differences in implementations.
So we need to check if it's
a directory and don't count
sum it's size to the whole if it is.