Install pip deps from the directory of environment.yml
Suppose we have the following dependencies in the environment.yml:

```
dependencies:
  - pip
  - pip:
    - -r requirements.txt
    - -e deps/my_local_package
```

Without this commit, it is impossible to install such dependencies with
micromamba. That is because micromamba was creating its own temporary
requirements.txt file in the /tmp directory and called pip to install them.

For some reason pip resolves `-r` paths as relative to the file that
contains them, but `-e` paths as relative to the current working directory.
It is a known issue in pip: https://github.com/pypa/pip/issues/8765.

So in the case of mamba, pip looks for `/tmp/requirements.txt` and
`directory/of/env/yml/deps/my_local_package` and unless whole project
is thrown right into `/tmp` this installation will fail.

This commit makes it so installing pip deps matches the behavior of conda.