Make xonsh completer only trigger when using micromamba
Currently, after installing `micromamba` and initializing it with the
`xonsh` shell, the completions for `micromamba` will show up upon use of
`<TAB>` for other executables, such as `rsync` or vim`.

This change ensures that `micromamba` completions do *not* appear for
cases in which `micromamba` is not the first argument to the context.

See for reference: https://github.com/xonsh/xonsh/issues/5692