Restore POSIX compatibility in micromamba profile script
-v in test seems to explicitly be _not_ POSIX compliant, and requires
bash v4+. This is a problem on macOS, where the system version of bash is <4
and usually the default shell is zsh now. This test works in zsh, so only
causes a problem when trying to source-activate the environment from a script
running inside /bin/bash.

This restores the POSIX compatible test, but also protects against opt -u by
defaulting to empty if the variable is not present.